requirements.txt
python-telegram-bot==13.15*
python-dotenv==1.2.1*
google-cloud-dialogflow==2.44.0*